+++ entity/src/email.rs
@@ -10,8 +10,8 @@ use validator::{Validate, ValidationError};
 #[sea_orm(table_name = "email")]
 pub struct Model {
     #[sea_orm(primary_key)]
-    pub id: i32,
-    pub user_id: i32,
+    pub id: i64,
+    pub user_id: i64,
     #[sea_orm(unique)]
     #[validate(email)]
     pub email: String,
