+++ entity/src/user.rs
@@ -10,7 +10,7 @@ use validator::{Validate, ValidationError};
 #[sea_orm(table_name = "user")]
 pub struct Model {
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     pub username: String,
     pub password: String,
     pub nickname: String,
@@ -20,17 +20,9 @@ pub struct Model {
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
 pub enum Relation {
-    // #[sea_orm(has_many = "super::author::Entity")]
-    // Author,
     #[sea_orm(has_many = "super::email::Entity")]
     Email,
 }
-
-// impl Related<super::author::Entity> for Entity {
-//     fn to() -> RelationDef {
-//         Relation::Author.def()
-//     }
-// }
 
 impl Related<super::email::Entity> for Entity {
     fn to() -> RelationDef {
